backend/requirements.txt
fastapi
uvicorn[standard]
# playwright  # archived — Unipile handles all LinkedIn actions; scraper/engine.py retained but dormant
supabase
cryptography
google-generativeai
yfinance
duckduckgo-search
python-dotenv
psycopg2-binary
requests
httpx>=0.27.0
